requirements.txt
GPUtil
IPython
Pillow
PyWavelets
graphviz
imageio
import_ipynb
keras==2.10.0
keras_sequential_ascii==0.1.1
# logging
matplotlib
nbimporter
netCDF4
numpy
opencv-pytho==4.7.0.72
packaging
pandas
pandasql
pathlib
# pgmagick
plotly
psutil
pynvml
pytest
# quiver_engine
regex
requests
scipy
seaborn
scikit-image
scikit-learn
split-folders[full]
sporco
tabulate
tensorboard==2.10.1
tensorflow==2.10.1
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm
typing
visualkeras
webcolors
